safe 14-01

+++ src/isometrie.c
@@ -17,13 +17,13 @@ static void	isometric(t_line *line)
 	t_point	new_start;
 	t_point	new_end;
 
-	new_start.x = (line->start.x - line->start.y) * cos(ANG_30);
-	new_start.y = (line->start.x + line->start.y) * sin(ANG_30) - \
+	new_start.x = (line->start.x - line->start.y) * cosf(ANG_30);
+	new_start.y = (line->start.x + line->start.y) * sinf(ANG_30) - \
 		line->start.z;
 	line->start.x = new_start.x;
 	line->start.y = new_start.y;
-	new_end.x = (line->end.x - line->end.y) * cos(ANG_30);
-	new_end.y = (line->end.x + line->end.y) * sin(ANG_30) - line->end.z;
+	new_end.x = (line->end.x - line->end.y) * cosf(ANG_30);
+	new_end.y = (line->end.x + line->end.y) * sinf(ANG_30) - line->end.z;
 	line->end.x = new_end.x;
 	line->end.y = new_end.y;
 }
