+++ src/rotate.c
@@ -17,12 +17,12 @@ void	rotate_x(t_line *line, double angle)
 	t_point	new_start;
 	t_point	new_end;
 
-	new_start.y = line->start.y * cos(angle) - line->start.z * sin(angle);
-	new_start.z = line->start.y * sin(angle) + line->start.z * cos(angle);
+	new_start.y = line->start.y * cosf(angle) - line->start.z * sinf(angle);
+	new_start.z = line->start.y * sinf(angle) + line->start.z * cosf(angle);
 	line->start.y = new_start.y;
 	line->start.z = new_start.z;
-	new_end.y = line->end.y * cos(angle) - line->end.z * sin(angle);
-	new_end.z = line->end.y * sin(angle) + line->end.z * cos(angle);
+	new_end.y = line->end.y * cosf(angle) - line->end.z * sinf(angle);
+	new_end.z = line->end.y * sinf(angle) + line->end.z * cosf(angle);
 	line->end.y = new_end.y;
 	line->end.z = new_end.z;
 }
@@ -32,12 +32,12 @@ void	rotate_y(t_line *line, double angle)
 	t_point	new_start;
 	t_point	new_end;
 
-	new_start.x = line->start.x * cos(angle) + line->start.z * sin(angle);
-	new_start.z = -line->start.x * sin(angle) + line->start.z * cos(angle);
+	new_start.x = line->start.x * cosf(angle) + line->start.z * sinf(angle);
+	new_start.z = -line->start.x * sinf(angle) + line->start.z * cosf(angle);
 	line->start.x = new_start.x;
 	line->start.z = new_start.z;
-	new_end.x = line->end.x * cos(angle) + line->end.z * sin(angle);
-	new_end.z = -line->end.x * sin(angle) + line->end.z * cos(angle);
+	new_end.x = line->end.x * cosf(angle) + line->end.z * sinf(angle);
+	new_end.z = -line->end.x * sinf(angle) + line->end.z * cosf(angle);
 	line->end.x = new_end.x;
 	line->end.z = new_end.z;
 }
@@ -47,12 +47,12 @@ void	rotate_z(t_line *line, double angle)
 	t_point	new_start;
 	t_point	new_end;
 
-	new_start.x = line->start.x * cos(angle) - line->start.y * sin(angle);
-	new_start.y = line->start.x * sin(angle) + line->start.y * cos(angle);
+	new_start.x = line->start.x * cosf(angle) - line->start.y * sinf(angle);
+	new_start.y = line->start.x * sinf(angle) + line->start.y * cosf(angle);
 	line->start.x = new_start.x;
 	line->start.y = new_start.y;
-	new_end.x = line->end.x * cos(angle) - line->end.y * sin(angle);
-	new_end.y = line->end.x * sin(angle) + line->end.y * cos(angle);
+	new_end.x = line->end.x * cosf(angle) - line->end.y * sinf(angle);
+	new_end.y = line->end.x * sinf(angle) + line->end.y * cosf(angle);
 	line->end.x = new_end.x;
 	line->end.y = new_end.y;
 }
